Move ParallelismInfo out of jobs into tool_util
This is part of #8240
(Figure out why packages/tool_util isn't being tested and fix the problems.)

This PR does not include updated documentation rst files. These will be
submitted as a separate PR due to a just discovered (potential) issue
with sphinx v. 1.8.5. (period is no longer escaped: foo.bar becomes
foo.bar)
